[gpu] Don't suppress size_t to int warnings in gpu/
BUG=167187

Cq-Include-Trybots: master.tryserver.chromium.android:android_optional_gpu_tests_rel;master.tryserver.chromium.linux:linux_optional_gpu_tests_rel;master.tryserver.chromium.mac:mac_optional_gpu_tests_rel;master.tryserver.chromium.win:win_optional_gpu_tests_rel

+++ gpu/config/gpu_info_collector_win.cc
@@ -24,6 +24,7 @@
 #include "base/message_loop/message_loop.h"
 #include "base/metrics/field_trial.h"
 #include "base/metrics/histogram_macros.h"
+#include "base/numerics/safe_conversions.h"
 #include "base/scoped_native_library.h"
 #include "base/strings/string16.h"
 #include "base/strings/string_number_conversions.h"
@@ -157,7 +158,7 @@ CollectInfoResult CollectDriverInfoD3D(const std::wstring& device_id,
           std::wstring id = new_device_id;
 
           if (id.compare(0, device_id.size(), device_id) == 0)
-            primary_device = drivers.size();
+            primary_device = base::saturated_cast<int>(drivers.size());
 
           uint32_t vendor_id = 0, device_id = 0;
           DeviceIDToVendorAndDevice(id, &vendor_id, &device_id);
